+++ components/src/main/java/com/mercadolibre/android/andesui/message/hierarchy/AndesMessageHierarchyInterface.kt
@@ -62,7 +62,7 @@ internal sealed class AndesMessageHierarchyInterface {
             buildColoredBitmapDrawable(
                     IconProvider(context).loadIcon("andes_ui_close_20") as BitmapDrawable,
                     context,
-                    hierarchy.dismissableIconColor()
+                    color = hierarchy.dismissableIconColor()
             )
 
     /**