+++ components/src/main/java/com/mercadolibre/android/andesui/textfield/AndesTextfield.kt
@@ -427,7 +427,7 @@ class AndesTextfield : ConstraintLayout {
         rightIcon.setImageDrawable(buildColoredBitmapDrawable(
                 IconProvider(context).loadIcon(iconPath) as BitmapDrawable,
                 context,
-                R.color.andes_gray_800.toAndesColor())
+                color = R.color.andes_gray_800.toAndesColor())
         )
     }
 
@@ -440,7 +440,7 @@ class AndesTextfield : ConstraintLayout {
         leftIcon.setImageDrawable(buildColoredBitmapDrawable(
                 IconProvider(context).loadIcon(iconPath) as BitmapDrawable,
                 context,
-                R.color.andes_gray_450.toAndesColor())
+                color = R.color.andes_gray_450.toAndesColor())
         )
     }